Conflicts:
	pyrem/task.py

+++ pyrem/task.py
@@ -12,12 +12,14 @@ import string
 
 from enum import Enum
 from subprocess import Popen, PIPE
-from threading import Thread
+from threading import RLock, Thread
+
+from utils import synchronized
 
 
 TaskStatus = Enum('TaskStatus', 'IDLE STARTED STOPPED')
 
-STARTED_TASKS = set() # TODO: figure out why some IDLE tasks end up here
+STARTED_TASKS = set()
 
 def cleanup():
     to_stop = STARTED_TASKS.copy()
@@ -30,15 +32,17 @@ atexit.register(cleanup)
 
 
 # TODO: create a wait_stopped() so that Tasks can be stopped in parallel
-
+# TODO: synchronizing Task methods might cause problems for the cleanup when
+#       the Task's methods are interrupted
 
 class Task(object):
-    # TODO: put a lock on Task objects to make all function calls atomic
 
     def __init__(self):
         self._status = TaskStatus.IDLE
+        self._lock = RLock()
         self.return_value = None # TODO: make this always a dict, rename return_values
 
+    @synchronized
     def start(self, wait=False):
         if self._status is not TaskStatus.IDLE:
             raise RuntimeError("Cannot start task in state %s" % self._status)
@@ -54,6 +58,7 @@ class Task(object):
     def _start(self):
         raise NotImplementedError
 
+    @synchronized
     def wait(self):
         if self._status is not TaskStatus.STARTED:
             raise RuntimeError("Cannot wait on task in state %s" % self._status)
@@ -64,6 +69,7 @@ class Task(object):
     def _wait(self):
         pass
 
+    @synchronized
     def stop(self):
         if self._status is TaskStatus.STOPPED:
             return
@@ -78,6 +84,7 @@ class Task(object):
     def _stop(self):
         pass
 
+    @synchronized
     def reset(self):
         if self._status is not TaskStatus.STOPPED:
             raise RuntimeError("Cannot reset task in state %s" % self._status)
@@ -92,16 +99,14 @@ class Task(object):
         return "Task(status=%s, return_value=%s)" % (
             self._status, self.return_value)
 
-# TODO: define a remote task that kills the remote processes started by it in
-#       stop
-
 
 class SubprocessTask(Task):
     DEVNULL = file(os.devnull, 'w')
 
     def __init__(self, command, quiet=False, return_output=False, shell=False):
         super(SubprocessTask, self).__init__()
-        self._command = command
+        assert isinstance(command, list)
+        self._command = [str(c) for c in command]
 
         self._popen_kwargs = {}
         self._popen_kwargs['stdin'] = self.DEVNULL
@@ -143,19 +148,24 @@ class SubprocessTask(Task):
 
 # TODO: add an option to skip the tmpfile, pgrep, and cleanup stuff
 class RemoteTask(SubprocessTask):
-    def __init__(self, host, command, quiet=False, return_output=False):
-        assert isinstance(command, list)
+    def __init__(self, host, command, quiet=False, return_output=False,
+                 kill_remote=True):
         self._host = host
-        # Temp file holds the PIDs of processes started on remote host
-        self._tmp_file_name = '/tmp/pyrem_procs-' + ''.join(
-            random.SystemRandom().choice(string.ascii_lowercase + string.digits)
-            for _ in range(8))
-        # TODO: Ending the user's command with ' & pgrep ...' might not be safe.
-        #       If the command ends in a &, for instance, this will just fail on
-        #       the spot. Try to figure out a good way around this, but at least
-        #       warn the user in RemoteTask's docstring
-        ssh_command = ['ssh', host, ' '.join(command) +
-                       ' & pgrep -P $$ >%s' % self._tmp_file_name]
+        assert isinstance(command, list)
+        ssh_command = ['ssh', host, ' '.join(command)]
+
+        self._kill_remote = kill_remote
+        if kill_remote:
+            # Temp file holds the PIDs of processes started on remote host
+            self._tmp_file_name = '/tmp/pyrem_procs-' + ''.join(
+                random.SystemRandom().choice(
+                    string.ascii_lowercase + string.digits)
+                for _ in range(8))
+            # TODO: Ending the user's command with ' & pgrep ...' might not be
+            # safe. If the command ends in a &, for instance, this will just
+            # fail on the spot. Try to figure out a good way around this, but at
+            # least warn the user in RemoteTask's docstring
+            ssh_command.append(' & pgrep -P $$ >%s' % self._tmp_file_name)
 
         super(RemoteTask, self).__init__(ssh_command, quiet=quiet,
                                          return_output=return_output,
@@ -164,10 +174,11 @@ class RemoteTask(SubprocessTask):
     def _stop(self):
         super(RemoteTask, self)._stop()
         # Silence the kill_proc to prevent messages about already killed procs
-        kill_proc = Popen(
-            ['ssh', self._host, 'kill -9 `cat %s`' % self._tmp_file_name],
-            stdout=self.DEVNULL, stderr=self.DEVNULL, stdin=self.DEVNULL)
-        kill_proc.wait()
+        if self._kill_remote:
+            kill_proc = Popen(
+                ['ssh', self._host, 'kill -9 `cat %s`' % self._tmp_file_name],
+                stdout=self.DEVNULL, stderr=self.DEVNULL, stdin=self.DEVNULL)
+            kill_proc.wait()
 
 
     def __repr__(self):
